NMEA Satellite Info parser: reject negative sentence numbers
When parsing an NMEA GSV message, we extract various integer numbers
that represent the total number of sentences, the index of the current
sentence, and the total number of visible satellites. These values
cannot be negative (and the first two also cannot be zero).
So, use quint32 to store these variables and use QBA::toUInt()
when reading them.

Since sentence and totalSats are now quint32, extend the range of
numSatInSentence to qint64, making sure that we do not accidentally
truncate the result.
This also requires to change the type of the index variable in the
for loop iterating over numSatInSentence.

Add unit-tests that cover these cases.

Since the very large negative sentence number could underflow when
multiplied by four, and thus cause an UB, pick this patch down to 6.5.
Do not pick to 5.15, because the entire satelliteinfo parser logic is
slightly different there, and also it is missing test coverage.

Task-number: QTBUG-146551
Pick-to: 6.11 6.8 6.5

+++ src/positioning/qlocationutils.cpp
@@ -397,19 +397,19 @@ QLocationUtils::getSatInfoFromNmea(QByteArrayView bv, QList<QGeoSatelliteInfo> &
         return QNmeaSatelliteInfoSource::FullyParsed; // Malformed sentence.
     }
     bool ok;
-    const int totalSentences = parts.at(1).toInt(&ok);
+    const quint32 totalSentences = parts.at(1).toUInt(&ok);
     if (!ok) {
         infos.clear();
         return QNmeaSatelliteInfoSource::FullyParsed; // Malformed sentence.
     }
 
-    const int sentence = parts.at(2).toInt(&ok);
+    const quint32 sentence = parts.at(2).toUInt(&ok);
     if (!ok) {
         infos.clear();
         return QNmeaSatelliteInfoSource::FullyParsed; // Malformed sentence.
     }
 
-    const int totalSats = parts.at(3).toInt(&ok);
+    const quint32 totalSats = parts.at(3).toUInt(&ok);
     if (!ok) {
         infos.clear();
         return QNmeaSatelliteInfoSource::FullyParsed; // Malformed sentence.
@@ -418,14 +418,14 @@ QLocationUtils::getSatInfoFromNmea(QByteArrayView bv, QList<QGeoSatelliteInfo> &
     if (sentence == 1)
         infos.clear();
 
-    const int numSatInSentence = qMin(sentence * 4, totalSats) - (sentence - 1) * 4;
+    const qint64 numSatInSentence = qMin(sentence * 4, totalSats) - (sentence - 1) * 4;
     if (parts.size() < (4 + numSatInSentence * 4)) {
         infos.clear();
         return QNmeaSatelliteInfoSource::FullyParsed; // Malformed sentence.
     }
 
     int field = 4;
-    for (int i = 0; i < numSatInSentence; ++i) {
+    for (qint64 i = 0; i < numSatInSentence; ++i) {
         QGeoSatelliteInfo info;
         info.setSatelliteSystem(system);
         int prn = parts.at(field++).toInt(&ok);
